minimum implementation of parser and validator

+++ include/deque.h
@@ -6,7 +6,7 @@
 /*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/09 23:12:30 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/07/10 02:07:39 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/07/11 14:14:42 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ typedef struct s_deque
     long long   sz;
 	long long	front;
 	long long	back;
-	int	*buffer;
+	long long	*buffer;
 }				t_deque;
 
 void	deque_init(t_deque *deque, long long sz);
@@ -30,7 +30,7 @@ bool	deque_is_empty(t_deque *deque);
 void	deque_free_buffer(t_deque *deque);
 void	deque_push_front(t_deque *deque, int data);
 void	deque_push_back(t_deque *deque, int data);
-int		*deque_pop_front(t_deque *deque);
-int		*deque_pop_back(t_deque *deque);
+long long		*deque_pop_front(t_deque *deque);
+long long		*deque_pop_back(t_deque *deque);
 
 #endif
